+++ src/pyodide/pyodide.ts
@@ -45,7 +45,20 @@ import { patchMatplotlib } from "./matplotlib";
    if (!baseURL.endsWith("/")) {
      baseURL += '/';
    }
- 
+
+   // Inject custom stdin/stdout/stderr
+   if(config.stdin) {
+    Module.preRun = [function() {
+      Module.FS.init(config.stdin, null, null);
+    }];
+  }
+  if(config.print) {
+    Module.print = config.print;
+  }
+  if(config.printErr) {
+    Module.printErr = config.printErr;
+  }
+
    ////////////////////////////////////////////////////////////
    // Package loading
    const DEFAULT_CHANNEL = "default channel";