+++ src/util.ts
@@ -1,3 +1,3 @@
-export function isPyProxy(val: any) {
-  return window.pyodide.isPyProxy(val);
+export function assertUnreachable(_x: never): never {
+  throw new Error("This case should have never been reached");
 }